Refactored Process into components + fixed Image Event payload bug
